refactored GameResets
